requirements.txt
pydantic
ruamel.yaml
behave
pytest